chore: reorganize agent directory
