+++ src/agent/core/AgentConfig.ts
