+++ src/agent/core/AgentDataclass.ts
