+++ src/agent/core/AgentState.ts
@@ -5,7 +5,7 @@
 // (none needed)
 
 // Local imports - log
-import * as logger from '../logger/logUtils';
+import * as logger from '../../logger/logUtils';
 
 // Local imports - agent components
 import {
