+++ src/agent/core/IAgent.ts
