+++ src/agent/core/ToolConfig.ts
