+++ src/agent/core/ToolState.ts
