+++ src/agent/modelHandlers/modelHandlerOpenAIResponse.ts
@@ -14,9 +14,9 @@ import type {
 
 // Local imports - base handler
 import { ModelHandlerOpenAI } from './modelHandlerOpenAI';
-import { ResponseUsageFactory } from '../ResponseUsage';
+import { ResponseUsageFactory } from '../core/ResponseUsage';
 import { calculateTokenPrice } from '../../utils/priceUtils';
-import { ToolState } from '../ToolState';
+import { ToolState } from '../core/ToolState';
 import { K_SLICE } from '../../utils/constants';
 
 // import { ResponseCreateParams } from 'openai/src/resources/responses/response';
