+++ src/agent/modelHandlers/modelHandlerOpenRouter.ts
@@ -6,7 +6,7 @@ import OpenAI from 'openai';
 
 // Local imports - agent components
 import { ModelHandlerOpenAI } from './modelHandlerOpenAI';
-import { ToolState } from '../ToolState';
+import { ToolState } from '../core/ToolState';
 import { K_SLICE } from '../../utils/constants';
 
 /**
