+++ src/agent/utils/userVars.ts
@@ -2,13 +2,13 @@
 
 import * as path from 'path';
 
-import { AgentLogger } from '../logger/AgentLogger';
-import { readFile } from '../utils/workspaceFileUtils';
-import { setVarFromFile } from '../utils/fileVarUtils';
-import { getXmlFormatFromFiles, getListOfFiles } from '../utils/promptUtils';
-import { AgentConfig } from './AgentConfig';
-import { AgentSetting } from './AgentDataclass';
-import { ModelHandler } from './modelHandlers';
+import { AgentLogger } from '../../logger/AgentLogger';
+import { readFile } from '../../utils/workspaceFileUtils';
+import { setVarFromFile } from '../../utils/fileVarUtils';
+import { getXmlFormatFromFiles, getListOfFiles } from '../../utils/promptUtils';
+import { AgentConfig } from '../core/AgentConfig';
+import { AgentSetting } from '../core/AgentDataclass';
+import { ModelHandler } from '../modelHandlers';
 
 /**
  * Build all user variables needed for prompt rendering.
