+++ src/webview/MessageHandler.ts
@@ -27,8 +27,8 @@ import { polishTextWithAI, FileContext } from '../utils/textEnhancementUtils';
 import { sleep } from '../utils/timeUtils';
 
 // Local imports - agent
-import { ToolConfig } from '../agent/ToolConfig';
-import { AgentConfig } from '../agent/AgentConfig';
+import { ToolConfig } from '../agent/core/ToolConfig';
+import { AgentConfig } from '../agent/core/AgentConfig';
 
 const CHANNEL = 'MessageHandler';
